requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf_yasg==1.21.10
Requests==2.32.4
